requirements.txt
fastapi
uvicorn[standard]
python-dotenv
things.py
